Read and Write FB - CS Core -WIP

+++ yabause/src/vdp1/vdp1_compute.h
@@ -18,7 +18,7 @@ extern void vdp1_compute_init(int width, int height, float ratiow, float ratioh)
 extern int* vdp1_compute(Vdp2 *varVdp2Regs, int id);
 extern int vdp1_add(vdp1cmd_struct* cmd, int clipcmd);
 extern void vdp1_clear(int id, float *col);
-extern u8* vdp1_get_directFB(Vdp2 *varVdp2Regs, int id);
+extern u32* vdp1_get_directFB(Vdp2 *varVdp2Regs, int id);
 extern void vdp1_set_directFB(int id);
 extern void vdp1_setup(void);
